requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cuda-toolkit==13.0.1
cycler==0.12.1
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.7
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.2
networkx==3.5
numpy==2.2.6
nvidia-cublas==13.0.0.19
nvidia-cuda-cupti==13.0.48
nvidia-cuda-nvrtc==13.0.48
nvidia-cuda-runtime==13.0.88
nvidia-cudnn-cu13==9.13.0.50
nvidia-cufft==12.0.0.15
nvidia-cufile==1.15.0.42
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.3.29
nvidia-cusparse==12.6.2.49
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.27.7
nvidia-nvjitlink==13.0.39
nvidia-nvshmem-cu13==3.3.24
nvidia-nvtx==13.0.39
onnx==1.19.1
onnxruntime-gpu==1.23.2
onnxslim==0.1.74
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.35.2
polars-runtime-32==1.35.2
protobuf==6.33.1
psutil==7.1.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
scipy==1.16.3
setuptools==70.2.0
six==1.17.0
sympy==1.14.0
tensorrt==10.14.1.48.post1
tensorrt_cu13==10.14.1.48.post1
tensorrt_cu13_bindings==10.14.1.48.post1
tensorrt_cu13_libs==10.14.1.48.post1
torch==2.9.1+cu130
torchaudio==2.9.1+cu130
torchvision==0.24.1+cu130
triton==3.5.1
typing_extensions==4.15.0
ultralytics==8.3.228
ultralytics-thop==2.0.18
urllib3==2.5.0
